pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "powerflow_mcp"
version = "0.1.0"
description = "MCP for running powerflow and analysis powerflow result for AI agent"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [
    {name = "Project Contributors"}
]
dependencies = [
    "pandapower",
    "numpy",
    "matplotlib",
    "mcp[cli]",
    "httpx",
    "openai",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "ruff",
]

[tool.ruff]
exclude = [
    ".git",
    ".ruff_cache",
    "build",
    "dist",
    ".venv",
]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "B"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
